feat: rename player by Course and create course module
