+++ src/App.tsx
@@ -1,14 +1,14 @@
 import "./styles/global.css";
 import { Provider as ReduxProvider } from "react-redux";
 import { store } from "./store";
-import { Player } from "./pages/Player";
+import { Course } from "./pages/Course.page";
 import { HelmetProvider } from "react-helmet-async";
 
 export function App() {
   return (
     <ReduxProvider store={store}>
       <HelmetProvider context={{}}>
-        <Player />
+        <Course />
       </HelmetProvider>
     </ReduxProvider>
   );
